+++ src/requests.rs
@@ -1,5 +1,5 @@
-use crate::channel::{Receiver, Sender};
 use actix_web::http::header::HeaderMap;
+use crossbeam::channel::{Receiver, Sender};
 use std::{collections::HashMap, rc::Rc};
 
 #[derive(Debug)]
@@ -44,11 +44,12 @@ impl ShareRequest {
 mod tests {
 
     use super::*;
+    use crossbeam::channel;
     use std::rc::Rc;
 
     #[test]
     fn request_receiver_is_empty() {
-        let (tx, rx) = crate::channel::unbounded();
+        let (tx, rx) = channel::unbounded();
         let rr = RequestReceiver { rx: Rc::new(rx) };
 
         assert!(rr.is_empty());
@@ -60,7 +61,7 @@ mod tests {
 
     #[test]
     fn request_reciever_len() {
-        let (tx, rx) = crate::channel::unbounded();
+        let (tx, rx) = channel::unbounded();
         let rr = RequestReceiver { rx: Rc::new(rx) };
 
         assert_eq!(rr.len(), 0);
@@ -72,7 +73,7 @@ mod tests {
 
     #[test]
     fn request_reciever_next() {
-        let (tx, rx) = crate::channel::unbounded();
+        let (tx, rx) = channel::unbounded();
         let rr = RequestReceiver { rx: Rc::new(rx) };
 
         assert!(rr.next().is_none());
@@ -82,7 +83,7 @@ mod tests {
         assert!(rr.next().is_some());
     }
 
-    fn add_request(tx: crate::channel::Sender<Request>) {
+    fn add_request(tx: channel::Sender<Request>) {
         if let Err(err) = tx.send(Request {
             body: String::new(),
             headers: HeaderMap::new(),